what i got so far from the first 10 minutes of a tutorial
added backend folder and  created package-json with mysql and express as dependencies. Also built a really simple (and quite ugly) input page for every table in the relational schema. So far, this does nothing as I have yet to get to the part in the tutorial which helps me connect the react project to the mySQL database (have installed mySQL so I have a database? not sure if its only local to my machine or not, also have yet to connect it really to react sorry)