Working on Connecting Database and FrontEnd
Added code in index.js (in the backend folder) that is a step towards sending data from the front end to the database. Basically formatting an insert query into the database with the variables that get filled out in App.js. Have yet to connect the 2 or test to see if it works.