Add Input to Table to Front End
The database is now connected to the SQLTable. And dubiously, records entered in the front end can be added! There is no constraints on what data can be entered however (e.g.  adding an EID that doesn't exist to a relation) and anything involving a checkbox currently throws an error. Will hopefully fix soon though